requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.6.0
kiwisolver==1.4.7
langchain==0.3.3
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.135
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
sequential==1.0.0
setuptools==75.1.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.35
sympy==1.13.3
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.1
-e git+https://github.com/TK17250/Uniqlo-WebScraping.git@e5a28052d4ce55f1b2b3c1b10d946e797d93c0a2#egg=uniqlowebscraping
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.15.2
